+++ backend/core/src/cache/types/redis-types.ts
@@ -1,0 +1,12 @@
+import { Store } from "cache-manager"
+import Redis from "redis"
+
+interface RedisStore extends Store {
+  name: "redis"
+  getClient: () => Redis.RedisClient
+  isCacheableValue: (value: unknown) => boolean
+}
+
+export interface RedisCache extends Cache {
+  store: RedisStore
+}