+++ sites/partners/src/listings/PaperListingForm/index.tsx
@@ -13,6 +13,11 @@ import {
   TimeFieldPeriod,
   Modal,
   AppearanceBorderType,
+  Tabs,
+  TabList,
+  Tab,
+  TabPanel,
+  LatitudeLongitude,
 } from "@bloom-housing/ui-components"
 import { useForm, FormProvider } from "react-hook-form"
 import {
@@ -180,11 +185,8 @@ const defaults: FormListing = {
   waitlistMaxSize: null,
   isWaitlistOpen: null,
   waitlistOpenSpots: null,
-  whatToExpect: {
-    applicantsWillBeContacted: "",
-    allInfoWillBeVerified: "",
-    bePreparedIfChosen: "",
-  },
+  whatToExpect:
+    "Applicants will be contacted by the property agent in rank order until vacancies are filled. All of the information that you have provided will be verified and your eligibility confirmed. Your application will be removed from the waitlist if you have made any fraudulent statements. If we cannot verify a housing preference that you have claimed, you will not receive the preference but will not be otherwise penalized. Should your application be chosen, be prepared to fill out a more detailed application and provide required supporting documents.",
   units: [],
   accessibility: "",
   amenities: "",
@@ -203,6 +205,7 @@ const defaults: FormListing = {
   urlSlug: undefined,
   showWaitlist: false,
   reviewOrderType: null,
+  unitsSummary: [],
   unitsSummarized: {
     unitTypes: [],
     priorityTypes: [],
@@ -232,7 +235,9 @@ const formatFormData = (
   units: TempUnit[],
   openHouseEvents: TempEvent[],
   preferences: Preference[],
-  paperApplications
+  paperApplications,
+  saveLatLong: LatitudeLongitude,
+  customPinPositionChosen: boolean
 ) => {
   const showWaitlistNumber =
     data.waitlistOpenQuestion === YesNoAnswer.Yes && data.waitlistSizeQuestion === YesNoAnswer.Yes
@@ -370,6 +375,12 @@ const formatFormData = (
     disableUnitsAccordion: stringToBoolean(data.disableUnitsAccordion),
     units: units,
     preferences: preferences,
+    buildingAddress: {
+      ...data.buildingAddress,
+      latitude: saveLatLong.latitude ?? null,
+      longitude: saveLatLong.longitude ?? null,
+    },
+    customMapPin: customPinPositionChosen,
     isWaitlistOpen: data.waitlistOpenQuestion === YesNoAnswer.Yes,
     applicationDueDate: applicationDueDateFormatted,
     yearBuilt: data.yearBuilt ? Number(data.yearBuilt) : null,
@@ -420,6 +431,7 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
 
   const { listingsService } = useContext(AuthContext)
 
+  const [tabIndex, setTabIndex] = useState(0)
   const [alert, setAlert] = useState<AlertErrorType | null>(null)
   const [loading, setLoading] = useState<boolean>(false)
   const [status, setStatus] = useState<ListingStatus>(null)
@@ -427,6 +439,19 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
   const [units, setUnits] = useState<TempUnit[]>([])
   const [openHouseEvents, setOpenHouseEvents] = useState<TempEvent[]>([])
   const [preferences, setPreferences] = useState<Preference[]>(listing?.preferences ?? [])
+  const [latLong, setLatLong] = useState<LatitudeLongitude>({
+    latitude: listing?.buildingAddress?.latitude ?? null,
+    longitude: listing?.buildingAddress?.longitude ?? null,
+  })
+  const [customMapPositionChosen, setCustomMapPositionChosen] = useState(
+    listing?.customMapPin || false
+  )
+
+  const setLatitudeLongitude = (latlong: LatitudeLongitude) => {
+    if (!loading) {
+      setLatLong(latlong)
+    }
+  }
 
   const [digitalApplicationMethod, setDigitalApplicationMethod] = useState<ApplicationMethod>(null)
   const [paperApplicationMethod, setPaperApplicationMethod] = useState<ApplicationMethod>(null)
@@ -537,10 +562,10 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
           units,
           openHouseEvents,
           orderedPreferences,
-          paperApplications
+          paperApplications,
+          latLong,
+          customMapPositionChosen
         )
-        console.info(formattedData)
-        //        throw new Error("whaaa!")
         const result = editMode
           ? await listingsService.update({
               listingId: listing.id,
@@ -571,6 +596,8 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
       router,
       preferences,
       paperApplications,
+      latLong,
+      customMapPositionChosen,
     ]
   )
 
@@ -594,7 +621,7 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
               <Button
                 inlineIcon="left"
                 icon="arrowBack"
-                onClick={() => (editMode ? router.push(`/listing/${listing?.id}`) : router.back())}
+                onClick={() => (editMode ? router.push(`/listings/${listing?.id}`) : router.back())}
               >
                 {t("t.back")}
               </Button>
@@ -627,32 +654,78 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
 
                 <Form id="listing-form" onSubmit={handleSubmit(triggerSubmit, onError)}>
                   <div className="flex flex-row flex-wrap">
-                    <div className="info-card md:w-9/12">
-                      <ListingIntro />
-                      <ListingPhoto />
-                      <BuildingDetails />
-                      <CommunityType listing={listing} />
-                      <Units
-                        units={units}
-                        setUnits={setUnits}
-                        disableUnitsAccordion={listing?.disableUnitsAccordion}
-                      />
-                      <Preferences preferences={preferences} setPreferences={setPreferences} />
-                      <AdditionalFees />
-                      <BuildingFeatures />
-                      <AdditionalEligibility />
-                      <AdditionalDetails />
-                      <RankingsAndResults listing={listing} />
-                      <LeasingAgent />
+                    <div className="md:w-9/12 pb-24">
+                      <Tabs
+                        forceRenderTabPanel={true}
+                        selectedIndex={tabIndex}
+                        onSelect={(index) => setTabIndex(index)}
+                      >
+                        <TabList>
+                          <Tab>Listing Details</Tab>
+                          <Tab>Application Process</Tab>
+                        </TabList>
+                        <TabPanel>
+                          <ListingIntro />
+                          <ListingPhoto />
+                          <BuildingDetails
+                            listing={listing}
+                            setLatLong={setLatitudeLongitude}
+                            latLong={latLong}
+                            customMapPositionChosen={customMapPositionChosen}
+                            setCustomMapPositionChosen={setCustomMapPositionChosen}
+                          />
+                          <CommunityType listing={listing} />
+                          <Units
+                            units={units}
+                            setUnits={setUnits}
+                            disableUnitsAccordion={listing?.disableUnitsAccordion}
+                          />
+                          <Preferences preferences={preferences} setPreferences={setPreferences} />
+                          <AdditionalFees />
+                          <BuildingFeatures />
+                          <AdditionalEligibility />
+                          <AdditionalDetails />
 
-                      <ApplicationTypes listing={listing} methodsState={methodsState} />
+                          <div className="text-right -mr-8 -mt-8 relative" style={{ top: "7rem" }}>
+                            <Button
+                              type="button"
+                              icon="arrowForward"
+                              onClick={() => {
+                                setTabIndex(1)
+                                window.scrollTo({ top: 0, behavior: "smooth" })
+                              }}
+                            >
+                              Application Process
+                            </Button>
+                          </div>
+                        </TabPanel>
+                        <TabPanel>
+                          <RankingsAndResults listing={listing} />
+                          <LeasingAgent />
+                          <ApplicationTypes listing={listing} methodsState={methodsState} />
+                          <ApplicationAddress listing={listing} />
+                          <ApplicationDates
+                            listing={listing}
+                            openHouseEvents={openHouseEvents}
+                            setOpenHouseEvents={setOpenHouseEvents}
+                          />
 
-                      <ApplicationAddress listing={listing} />
-                      <ApplicationDates
-                        listing={listing}
-                        openHouseEvents={openHouseEvents}
-                        setOpenHouseEvents={setOpenHouseEvents}
-                      />
+                          <div className="-ml-8 -mt-8 relative" style={{ top: "7rem" }}>
+                            <Button
+                              type="button"
+                              icon="arrowBack"
+                              iconPlacement="left"
+                              onClick={() => {
+                                setTabIndex(0)
+                                window.scrollTo({ top: 0, behavior: "smooth" })
+                              }}
+                            >
+                              Listing Details
+                            </Button>
+                          </div>
+                        </TabPanel>
+                      </Tabs>
+
                       {listing?.status === ListingStatus.closed && (
                         <LotteryResults
                           submitCallback={(data) => {